fix(runtime): preserve audit closure across repair loops
