+++ .feature-specs/SKILL-135-audit-first-review-gate/decomposition-manifest.yaml
@@ -3,24 +3,26 @@ contract_version: "0.5"
 issue_key: "SKILL-135"
 feature_name: "audit-first-review-gate"
 parent_spec_path: ".feature-specs/SKILL-135-audit-first-review-gate/spec.md"
-status: "in_progress"
+status: "blocked"
 execution_model: "same_branch_commit_per_subtask"
 base_branch: "main"
 feature_branch: "feat/SKILL-135-audit-first-review-gate"
 stack_branches: []
 current_subtask_intent:
   subtask_id: 1
-  action: "start"
+  action: "blocked"
 subtasks:
 - id: 1
   name: "Audit-first phase reordering"
   spec_path: ".feature-specs/SKILL-135-audit-first-review-gate/spec_subtask_1_phase_reordering.md"
-  status: "pending"
+  status: "blocked"
   branch: null
   commit_sha: null
   workflow_id: "wftr-20260720-153328-g6un"
-  blocked_reason: null
-  last_resumable_step: null
+  blocked_reason: "Phase output reported status 'blocked'. Committed the validated\
+    \ changes locally, but push was blocked by a divergent remote branch; safely aborted\
+    \ the conflicting merge without altering the commit."
+  last_resumable_step: "commit_push"
   linear_issue_id: null
   finalizing_agent_id: null
   participating_agent_ids: []
