+++ runtime-kotlin/runtime-application/src/main/kotlin/skillbill/application/featuretask/FeatureTaskRuntimePhaseBriefingAssembler.kt
@@ -138,21 +138,7 @@ object FeatureTaskRuntimePhaseBriefingAssembler {
       FeatureTaskRuntimePhaseWorkflowDefinition.ceremonyScaling(invariants.featureSize)
         .toBriefingLines()
         .forEach { line -> appendLine("  $line") }
-      appendLine("acceptance_criteria:")
-      // The runtime owns the AC-NNN ref so the audit cannot invent a competing ordinal convention and
-      // durable closure cannot be keyed on a ref no briefing ever emitted.
-      val closedCriterionRefs = handoff.durablyClosedCriterionRefs.toSet()
-      invariants.acceptanceCriteria.forEachIndexed { index, criterion ->
-        val criterionRef = canonicalAcceptanceCriterionRef(index + 1)
-        if (criterionRef !in closedCriterionRefs) {
-          appendLine("  $criterionRef. $criterion")
-        }
-      }
-      if (closedCriterionRefs.isNotEmpty()) {
-        appendLine("durably_closed_criteria:")
-        appendLine("  (each reached a satisfied verdict and is closed; do not re-verify or report a gap against it)")
-        closedCriterionRefs.sorted().forEach { criterionRef -> appendLine("  - $criterionRef") }
-      }
+      appendAcceptanceCriteria(handoff)
       appendLine("mandates_and_overrides:")
       if (invariants.mandatesAndOverrides.isEmpty()) {
         appendLine("  (none)")
@@ -177,6 +163,20 @@ object FeatureTaskRuntimePhaseBriefingAssembler {
         append(handoff.derivedContextKeys.joinToString(separator = "\n") { key -> "- $key" })
       }
     }
+
+  private fun StringBuilder.appendAcceptanceCriteria(handoff: FeatureTaskRuntimePhaseHandoff) {
+    appendLine("acceptance_criteria:")
+    val closedCriterionRefs = handoff.durablyClosedCriterionRefs.toSet()
+    handoff.runInvariants.acceptanceCriteria.forEachIndexed { index, criterion ->
+      val criterionRef = canonicalAcceptanceCriterionRef(index + 1)
+      if (criterionRef !in closedCriterionRefs) appendLine("  $criterionRef. $criterion")
+    }
+    if (closedCriterionRefs.isNotEmpty()) {
+      appendLine("durably_closed_criteria:")
+      appendLine("  (each reached a satisfied verdict and is closed; do not re-verify or report a gap against it)")
+      closedCriterionRefs.sorted().forEach { criterionRef -> appendLine("  - $criterionRef") }
+    }
+  }
 
   /**
    * Bounds one inlined upstream body to [budgetBytes] UTF-8 bytes, returning it verbatim when it
