+++ runtime-kotlin/runtime-application/src/main/kotlin/skillbill/application/featuretask/FeatureTaskRuntimePhaseRecorder.kt
@@ -28,18 +28,12 @@ import skillbill.workflow.taskruntime.model.FEATURE_TASK_RUNTIME_PHASE_LEDGER_LI
 import skillbill.workflow.taskruntime.model.FEATURE_TASK_RUNTIME_PHASE_RECORDS_ARTIFACT_KEY
 import skillbill.workflow.taskruntime.model.FEATURE_TASK_RUNTIME_PHASE_STATUS_BLOCKED
 import skillbill.workflow.taskruntime.model.FEATURE_TASK_RUNTIME_RESOLVED_BRANCH_ARTIFACT_KEY
-import skillbill.workflow.taskruntime.model.FeatureTaskRuntimeAuditGap
-import skillbill.workflow.taskruntime.model.FeatureTaskRuntimeAuditRepairPlan
-import skillbill.workflow.taskruntime.model.FeatureTaskRuntimeAuditRepairProgress
 import skillbill.workflow.taskruntime.model.FeatureTaskRuntimeAuditRepairState
 import skillbill.workflow.taskruntime.model.FeatureTaskRuntimeDecomposeTerminal
 import skillbill.workflow.taskruntime.model.FeatureTaskRuntimePhaseLedgerEntry
 import skillbill.workflow.taskruntime.model.FeatureTaskRuntimePhaseRecord
-import skillbill.workflow.taskruntime.model.FeatureTaskRuntimePriorGapDisposition
 import skillbill.workflow.taskruntime.model.FeatureTaskRuntimeRepairItemResult
 import skillbill.workflow.taskruntime.model.FeatureTaskRuntimeResolvedBranch
-import skillbill.workflow.taskruntime.model.FeatureTaskRuntimeUnresolvedGap
-import skillbill.workflow.taskruntime.model.FeatureTaskRuntimeUnresolvedGapLedger
 import skillbill.workflow.taskruntime.model.FeatureTaskRuntimeVerdict
 import skillbill.workflow.taskruntime.model.GOAL_SUBTASK_REVIEW_RESULTS_ARTIFACT_KEY
 import skillbill.workflow.taskruntime.model.GOAL_SUBTASK_REVIEW_STATE_ARTIFACT_KEY
@@ -152,7 +146,7 @@ class FeatureTaskRuntimePhaseRecorder(
       ) {
         mapOf(
           FEATURE_TASK_RUNTIME_AUDIT_REPAIR_STATE_ARTIFACT_KEY to auditRepairStateToWire(
-            reconcileAuditRepairState(
+            FeatureTaskRuntimeAuditRepairReconciler.reconcile(
               AuditRepairReconciliation(
                 prior = priorAuditState,
                 latestPlan = latestPlan,
@@ -181,205 +175,6 @@ class FeatureTaskRuntimePhaseRecorder(
       )
       true
     }
-  }
-
-  private data class AuditRepairReconciliation(
-    val prior: FeatureTaskRuntimeAuditRepairState?,
-    val latestPlan: FeatureTaskRuntimeAuditRepairPlan?,
-    val repairResults: List<FeatureTaskRuntimeRepairItemResult>,
-    val dispositions: List<FeatureTaskRuntimePriorGapDisposition>?,
-    val repositoryFingerprint: String?,
-    val edgeIteration: Int?,
-    val auditWrite: Boolean = false,
-    val auditScopeCriterionRefs: List<String> = emptyList(),
-  )
-
-  private data class GapReconciliation(
-    val dispositions: List<FeatureTaskRuntimePriorGapDisposition>,
-    val recurringIds: Set<String>,
-    val latestIds: Set<String>,
-    val unresolvedGaps: List<FeatureTaskRuntimeUnresolvedGap>,
-    val closedGenerationHighWaterMarks: Map<String, Int>,
-  )
-
-  private fun reconcileAuditRepairState(input: AuditRepairReconciliation): FeatureTaskRuntimeAuditRepairState {
-    val acceptedPlans = listOfNotNull(input.latestPlan ?: input.prior?.acceptedPlans?.lastOrNull())
-    if (acceptedPlans.isEmpty()) schemaError("Audit-repair state requires an accepted plan.")
-    val gaps = reconcileUnresolvedGaps(input)
-    val latestItemIds = acceptedPlans.single().gaps
-      .flatMap { it.repairItems }
-      .mapTo(linkedSetOf()) { it.repairItemId }
-    val allResults = reconcileLatestRepairResults(
-      input.prior?.repairItemResults.orEmpty(),
-      input.repairResults,
-      latestItemIds,
-    )
-    val progress = reconcileProgress(input, gaps, latestItemIds.size)
-    // Every other reconciliation failure surfaces as a typed workflow-state error; a ledger-cap or
-    // counter-coherence violation reaching the caller as a bare argument exception would durably wedge the
-    // workflow with an untyped message, so the write seam validates here rather than only on the next read.
-    return runCatching {
-      FeatureTaskRuntimeAuditRepairState(
-        acceptedPlans = acceptedPlans,
-        repairItemResults = allResults,
-        priorGapDispositions = gaps.dispositions,
-        unresolvedGapLedger = FeatureTaskRuntimeUnresolvedGapLedger(
-          gaps.unresolvedGaps,
-          gaps.closedGenerationHighWaterMarks,
-        ),
-        repositoryFingerprint = input.repositoryFingerprint ?: input.prior?.repositoryFingerprint,
-        progress = progress,
-        satisfiedCriterionRefs = reconcileSatisfiedCriteria(input, gaps),
-      ).also { it.requireDurableCoherence() }
-    }.getOrElse { error ->
-      if (error is IllegalArgumentException) {
-        schemaError("Audit-repair state is incoherent and was not persisted: ${error.message.orEmpty()}")
-      } else {
-        throw error
-      }
-    }
-  }
-
-  /**
-   * Durable closure of the criteria this audit verified and did not report a gap against. It is
-   * derived from absence-of-gap in a per-criterion audit intersected with the unresolved-gap ledger,
-   * so no schema bump is needed and closure stays verifiable from durable state. The union keeps it
-   * strictly monotone: a criterion once closed is never reopened, which is what bounds the otherwise
-   * uncapped audit_gap cycle by shrinking the unsatisfied set on every iteration.
-   */
-  private fun reconcileSatisfiedCriteria(
-    input: AuditRepairReconciliation,
-    gaps: GapReconciliation,
-  ): List<String> {
-    val prior = input.prior?.satisfiedCriterionRefs.orEmpty()
-    if (!input.auditWrite) return prior
-    val reportedCriteria = input.latestPlan?.gaps.orEmpty().mapTo(linkedSetOf()) { it.acceptanceCriterionRef }
-    // A criterion whose gap is still unresolved is never closed, so the durable-ledger gate that
-    // requires the next audit to disposition every unresolved gap stays satisfiable.
-    val stillUnresolved = gaps.unresolvedGaps.mapTo(linkedSetOf()) { it.acceptanceCriterionRef }
-    val newlyClosed = input.auditScopeCriterionRefs
-      .filterNot { it in reportedCriteria || it in stillUnresolved }
-    return (prior + newlyClosed).distinct()
-  }
-
-  private fun reconcileProgress(
-    input: AuditRepairReconciliation,
-    gaps: GapReconciliation,
-    latestPlanItemCount: Int,
-  ): FeatureTaskRuntimeAuditRepairProgress {
-    // Attempted counts every item the remediation phase was handed; resolved counts the terminal
-    // results it actually returned. They diverge when a phase reports results for only part of its plan.
-    val newlyAttemptedCount = if (input.repairResults.isEmpty()) 0 else latestPlanItemCount
-    val priorPlanGapIds = input.prior?.unresolvedGapLedger?.unresolvedGaps.orEmpty().mapTo(linkedSetOf()) { it.gapId }
-    // Only an audit write observes a fresh gap set; recomputing on the remediation write would compare
-    // the audit's own already-persisted ledger against itself and clobber both counters to zero.
-    val auditWrite = input.latestPlan != null
-    val ledgerSize = gaps.unresolvedGaps.size
-    val recurringGapCount = if (auditWrite) {
-      gaps.recurringIds.size
-    } else {
-      (input.prior?.progress?.recurringGapCount ?: 0).coerceAtMost(ledgerSize)
-    }
-    val newGapCount = if (auditWrite) {
-      gaps.latestIds.count { it !in priorPlanGapIds }
-    } else {
-      (input.prior?.progress?.newGapCount ?: 0).coerceAtMost(ledgerSize - recurringGapCount)
-    }
-    return FeatureTaskRuntimeAuditRepairProgress(
-      firstPassConvergence = false,
-      recurringGapCount = recurringGapCount,
-      newGapCount = newGapCount,
-      attemptedRepairItemCount = (input.prior?.progress?.attemptedRepairItemCount ?: 0) + newlyAttemptedCount,
-      resolvedRepairItemCount = (input.prior?.progress?.resolvedRepairItemCount ?: 0) + input.repairResults.size,
-      auditGapIterationCount = maxOf(
-        input.prior?.progress?.auditGapIterationCount ?: 0,
-        input.edgeIteration ?: 0,
-      ),
-    )
-  }
-
-  private fun reconcileUnresolvedGaps(input: AuditRepairReconciliation): GapReconciliation {
-    val priorLedger = input.prior?.unresolvedGapLedger
-    val priorUnresolved = priorLedger?.unresolvedGaps.orEmpty()
-    val priorIds = priorUnresolved.mapTo(linkedSetOf()) { it.gapId }
-    val dispositions = input.dispositions ?: input.prior?.priorGapDispositions.orEmpty()
-    if (input.dispositions != null && dispositions.mapTo(linkedSetOf()) { it.gapId } != priorIds) {
-      schemaError(
-        "Audit reconciliation must disposition every prior unresolved gap exactly once and cannot " +
-          "disposition a gap the durable ledger never carried; ledger=${priorIds.sorted()} " +
-          "dispositioned=${dispositions.map { it.gapId }.sorted()}.",
-      )
-    }
-    val resolvedIds = dispositions
-      .filter { it.status == FeatureTaskRuntimePriorGapDisposition.Status.RESOLVED }
-      .mapTo(linkedSetOf()) { it.gapId }
-    val recurringIds = dispositions
-      .filter { it.status == FeatureTaskRuntimePriorGapDisposition.Status.RECURRING }
-      .mapTo(linkedSetOf()) { it.gapId }
-    val latestGaps = input.latestPlan?.gaps
-      ?: if (input.repairResults.isNotEmpty()) input.prior?.acceptedPlans?.lastOrNull()?.gaps.orEmpty() else emptyList()
-    val latestIds = latestGaps.mapTo(linkedSetOf()) { it.gapId }
-    val criterionByPriorGapId = priorUnresolved.associate { it.gapId to it.acceptanceCriterionRef }
-    val resolvedCriteria = resolvedIds.mapNotNullTo(linkedSetOf(), criterionByPriorGapId::get)
-    val reopenedCriteria = latestGaps.map { it.acceptanceCriterionRef }.filter { it in resolvedCriteria }.sorted()
-    val durablyClosed = input.prior?.satisfiedCriterionRefs.orEmpty().toSet()
-    val reopenedClosedCriteria = latestGaps.map { it.acceptanceCriterionRef }
-      .filter { it in durablyClosed }
-      .sorted()
-    if (reopenedClosedCriteria.isNotEmpty()) {
-      schemaError(
-        "An audit cannot report a gap against an acceptance criterion that already reached a satisfied " +
-          "verdict and is durably closed; reopened criteria were $reopenedClosedCriteria.",
-      )
-    }
-    if (!latestIds.containsAll(recurringIds) || latestIds.any(resolvedIds::contains) ||
-      reopenedCriteria.isNotEmpty()
-    ) {
-      schemaError(
-        "Recurring gaps must retain their identities and a criterion dispositioned resolved cannot be " +
-          "re-reported by the same handoff under any gap id; reopened criteria were $reopenedCriteria.",
-      )
-    }
-    // Closing a gap consumes its generation. Carrying that forward is what makes the next identifier for a
-    // criterion derivable from durable state instead of supplied by the agent.
-    val surviving = priorUnresolved.filterNot { it.gapId in resolvedIds }
-    val marks = LinkedHashMap(priorLedger?.closedGenerationHighWaterMarks.orEmpty())
-    priorUnresolved.filter { it.gapId in resolvedIds }.forEach { gap ->
-      marks[gap.acceptanceCriterionRef] = maxOf(marks[gap.acceptanceCriterionRef] ?: 0, gap.generation)
-    }
-    return GapReconciliation(
-      dispositions = dispositions,
-      recurringIds = recurringIds,
-      latestIds = latestIds,
-      unresolvedGaps = mergeUnresolvedGaps(surviving, marks, latestGaps),
-      closedGenerationHighWaterMarks = marks,
-    )
-  }
-
-  private fun mergeUnresolvedGaps(
-    surviving: List<FeatureTaskRuntimeUnresolvedGap>,
-    closedGenerationHighWaterMarks: Map<String, Int>,
-    latestGaps: List<FeatureTaskRuntimeAuditGap>,
-  ): List<FeatureTaskRuntimeUnresolvedGap> {
-    val survivingLedger = FeatureTaskRuntimeUnresolvedGapLedger(surviving, closedGenerationHighWaterMarks)
-    val merged = linkedMapOf<String, FeatureTaskRuntimeUnresolvedGap>()
-    surviving.forEach { merged[it.gapId] = it }
-    latestGaps.forEach { gap ->
-      val derivedGapId = survivingLedger.allocateGapId(gap.acceptanceCriterionRef)
-      if (gap.gapId != derivedGapId) {
-        schemaError(
-          "gap_id '${gap.gapId}' for '${gap.acceptanceCriterionRef}' is not the identifier the runtime " +
-            "derives from durable state; expected '$derivedGapId'.",
-        )
-      }
-      merged[gap.gapId] = FeatureTaskRuntimeUnresolvedGap(
-        gapId = gap.gapId,
-        acceptanceCriterionRef = gap.acceptanceCriterionRef,
-        generation = gap.gapId.substringAfterLast('-').toIntOrNull()
-          ?: schemaError("Gap '${gap.gapId}' has no numeric generation."),
-      )
-    }
-    return merged.values.toList()
   }
 
   internal fun persistAuditGateInvalidation(workflowId: String, dbOverride: String? = null): Boolean =
